fix(types): major mypy type annotation improvements across Lambda functions
- Add proper type annotations to admin function parameters (AWS clients, loggers)
- Fix transcribe function duration calculation type (int vs float)
- Update process-manual function return type to Optional[str]
- Resolve query function parameter scoping by adding missing parameters
- Fixed 65% of mypy errors (630 → 222 errors)
- 7 out of 10 Lambda functions now pass mypy type checking
- Auto-format code with black and isort

+++ backend/src/functions/transcribe/app.py
@@ -133,7 +133,7 @@ def lambda_handler(event: Dict[str, Any], context: Any) -> Dict[str, Any]:
             cleanup_audio_file(s3_key, health_checker)
 
             # Calculate final costs
-            cost_params["lambda_duration_ms"] = (time.time() - start_time) * 1000
+            cost_params["lambda_duration_ms"] = int((time.time() - start_time) * 1000)
 
             request_cost = cost_calculator.calculate_request_cost(
                 context.aws_request_id if context else "unknown",
